migrate + renamed dotnet -> dotnet-version
