+++ lib/modules/datasource/dotnet-version/__fixtures__/releases-3.1.json
