+++ lib/modules/datasource/dotnet-version/__fixtures__/releases-5.0.json
