+++ lib/modules/datasource/dotnet-version/__fixtures__/releases-6.0.json
