+++ lib/modules/datasource/dotnet-version/__fixtures__/releases-7.0.json
