+++ lib/modules/datasource/dotnet-version/__fixtures__/releases-index.json
