+++ lib/modules/datasource/dotnet-version/schema.ts
